football/requirements.txt
flask
SQLAlchemy
Werkzeug==2.2.2
pandas
requests
mysql-connector-python
cassandra-driver
scikit-learn
dill